Split Interactivty and Advanced Features Colabs
